- Rename folder component to components - Create file for test Header
