+++ jest.config.ts
@@ -1,5 +1,6 @@
 import type { Config } from "jest";
 import nextJest from "next/jest.js";
+import "@testing-library/jest-dom";
 
 const createJestConfig = nextJest({
   // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
